// src/lib/types.ts
export type Product = {
  slug: string;
  featuredImage: { url: string };
  title: string;
  priceRange: { maxVariantPrice: { amount: string; currencyCode: string } };
}